# backend/cmd/app/docs/swagger.yaml
basePath: /api
definitions:
...
  data.Category:
    properties:
      id:
        type: string
      name:
        type: string
      storeId:
        type: string
    type: object
...
  data.Error:
    properties:
      error:
        type: string
    type: object
  data.Item:
    properties:
      category:
        type: string
      categoryId:
        type: string
      createdAt:
        type: string
      currentInventory:
        type: integer
      desc:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      price:
        type: number
      status:
        description: active, deleted, inactive
        type: string
      storeId:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: Boiboi API Documentation.
  title: Boiboi Backend API
  version: "1.0"
paths:
...
  /inventories/categories/:
    post:
      consumes:
      - application/json
      description: Create a new category for a store
      parameters:
      - description: Category object
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/data.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/data.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.Error'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Inventories
  /inventories/items/:
    post:
      consumes:
      - application/json
      description: Add a new item to a store's inventory
      parameters:
      - description: Item object
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/data.Item'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/data.Item'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/data.Error'
      security:
      - BearerAuth: []
      summary: Add item to store inventory
      tags:
      - Inventories
  /inventories/items/{id}:
    delete:
      description: Soft delete an item from a store's inventory by ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/data.Error'
      security:
      - BearerAuth: []
      summary: Remove item from store inventory
      tags:
      - Inventories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"